useNotifications hook to request notification permissions, get expo push token (#24)
Adds a useNotifications handler to request user permissions for notifications and manage user expoPushToken. Also adds listeners for when notifications are received. For now, the user token is printed to the console and is accessible in App.tsx.
Eventually, this should be integrated to the backend to update the cached user expoPushToken in mongoose when a new token is granted.

Co-authored-by: Michael Chen <michael.ld.chen@gmail.com>