Fix SignUp Bug
Before when a new user would sign up and be redirected to the onboarding, their jwt token would be stored
in the state such that when they exited the app and came back to sign in again, they would conintue to be
met with the onboarding form.

To resolve this, I added a button in the first onboarding form that would log the user logout and therefore
reset their state such that when they sign in again, they can login or sign with a different account.